+++ Guitar++/CLog.h
@@ -101,7 +101,7 @@ public:
 							i = ti - 1;
 							if (numbuf.size() > 0)
 							{
-								int argId = std::stoi(numbuf);
+								size_t argId = std::stoul(numbuf);
 
 								if (argId >= 0 && argId < a.size())
 								{
